fix async error
