+++ packages/ui/src/utils/errors.ts
@@ -1,0 +1,17 @@
+export function getErrorMessage(response: any) {
+  if (response.message) {
+    if (Array.isArray(response.message)) {
+      return formatErrorMessage(response.message[0]);
+    }
+    return formatErrorMessage(response.message);
+  }
+  return 'Unknown error occured.';
+}
+
+function formatErrorMessage(message: string) {
+  return message.charAt(0).toUpperCase() + message.slice(1);
+}
+
+export interface FormErrors {
+  error: string;
+}
